feat(asset-registry): add asset details, loading skeleton, history, purchase, technical details, and timeline components
- Implemented AssetDetailsTab component for displaying asset information, including image, status, and maintenance records.
- Created AssetLoadingSkeleton component for loading state representation.
- Developed HistoryTab component to show asset history events with appropriate icons and descriptions.
- Added PurchaseDetailsTab component for displaying purchase information and vendor details.
- Introduced TechnicalDetailsTab component for showcasing asset specifications and notes.
- Implemented TimelineItem component for rendering individual timeline events with icons and colors.
- Added getAuthenticatedUser function for retrieving user session and permissions.
- Created asset-details-repo for managing asset data retrieval and history.
- Established asset registration validation schema for consistent asset creation.
