+++ web/src/components/assets/office-electronics-registration-form.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { Filter, Paperclip, Plus, Search } from 'lucide-react';
 import Image from 'next/image';
 
-import { registerOfficeElectronicsAsset } from '@/actions/assets';
+import { registerAsset } from '@/actions/assets';
 import { useOpenRegistrationPanel } from '@/components/assets/use-open-registration-panel';
 import { FormPanel } from '@/components/shared/slide-panels/form-panel';
 import { tiqriToast } from '@/components/shared/sonner';
@@ -14,9 +14,9 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { SearchableDropdown } from '@/components/ui/searchable-dropdown';
 import {
-  initialRegisterOfficeElectronicsAssetActionState,
-  type RegisterOfficeElectronicsAssetActionState,
-} from '@/validations/office-electronics-asset';
+  initialRegisterAssetActionState,
+  type RegisterAssetActionState,
+} from '@/lib/validations/asset-registration';
 
 import {
   Table,
@@ -77,8 +77,8 @@ function getTodayDateValue() {
 }
 
 function getError(
-  state: RegisterOfficeElectronicsAssetActionState,
-  key: keyof NonNullable<RegisterOfficeElectronicsAssetActionState['errors']>
+  state: RegisterAssetActionState,
+  key: keyof NonNullable<RegisterAssetActionState['errors']>
 ) {
   return state.errors?.[key]?.[0];
 }
@@ -204,8 +204,8 @@ export function OfficeElectronicsRegistryPageClient({
   const openRegistrationPanel = useOpenRegistrationPanel(setIsPanelOpen);
   const invoiceInputRef = React.useRef<HTMLInputElement>(null);
   const [state, formAction, isPending] = React.useActionState(
-    registerOfficeElectronicsAsset,
-    initialRegisterOfficeElectronicsAssetActionState
+    registerAsset,
+    initialRegisterAssetActionState
   );
 
   const [categoryId, setCategoryId] = React.useState('');
