+++ web/src/components/features/asset-registry/panels/asset-loading-skeleton.tsx
